# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SwiftShip Logistics Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: riderId
          in: query
          schema:
            type: integer
        - name: stationId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: duplicateOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/bulk:
    post:
      operationId: createBulkOrders
      tags: [orders]
      summary: Create orders in bulk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkOrderBody"
      responses:
        "201":
          description: Bulk order result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkOrderResult"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with comments and history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/{id}/assign:
    post:
      operationId: assignOrder
      tags: [orders]
      summary: Assign or reassign a rider to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignOrderBody"
      responses:
        "200":
          description: Order assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    post:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/duplicate-action:
    post:
      operationId: handleDuplicateOrder
      tags: [orders]
      summary: Handle duplicate order (approve/reject/merge)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicateActionBody"
      responses:
        "200":
          description: Action taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Comments
  /orders/{orderId}/comments:
    get:
      operationId: listOrderComments
      tags: [comments]
      summary: List comments for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addOrderComment
      tags: [comments]
      summary: Add a comment to an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # Stock
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderCode:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        vendorCode:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        alternatePhone:
          type: ["string", "null"]
        productName:
          type: string
        productSku:
          type: ["string", "null"]
        quantity:
          type: integer
        codAmount:
          type: number
        deliveryCharge:
          type: number
        vendorPayable:
          type: number
        address:
          type: string
        landmark:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        stationId:
          type: ["integer", "null"]
        stationName:
          type: ["string", "null"]
        riderId:
          type: ["integer", "null"]
        riderName:
          type: ["string", "null"]
        priority:
          type: string
          enum: [normal, high, urgent]
        requestedDeliveryTime:
          type: ["string", "null"]
        status:
          type: string
        duplicateFlag:
          type: boolean
        duplicateReason:
          type: ["string", "null"]
        matchedOrderId:
          type: ["integer", "null"]
        duplicateConfidence:
          type: ["string", "null"]
        paymentReleaseStatus:
          type: string
          enum: [pending, released]
        notes:
          type: ["string", "null"]
        internalNote:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        deliveredAt:
          type: ["string", "null"]
      required:
        - id
        - orderCode
        - vendorId
        - vendorName
        - vendorCode
        - customerName
        - customerPhone
        - productName
        - quantity
        - codAmount
        - deliveryCharge
        - vendorPayable
        - address
        - priority
        - status
        - duplicateFlag
        - paymentReleaseStatus
        - createdAt
        - updatedAt
    OrderDetail:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
      required:
        - order
        - comments
        - statusHistory
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - orders
        - total
        - page
        - limit
    CreateOrderBody:
      type: object
      properties:
        vendorId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        alternatePhone:
          type: ["string", "null"]
        productName:
          type: string
        productSku:
          type: ["string", "null"]
        quantity:
          type: integer
        codAmount:
          type: number
        address:
          type: string
        landmark:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        stationId:
          type: ["integer", "null"]
        priority:
          type: string
        requestedDeliveryTime:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - vendorId
        - customerName
        - customerPhone
        - productName
        - quantity
        - codAmount
        - address
    UpdateOrderBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        alternatePhone:
          type: ["string", "null"]
        productName:
          type: string
        quantity:
          type: integer
        codAmount:
          type: number
        address:
          type: string
        landmark:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        stationId:
          type: ["integer", "null"]
        priority:
          type: string
        notes:
          type: ["string", "null"]
        internalNote:
          type: ["string", "null"]
    AssignOrderBody:
      type: object
      properties:
        riderId:
          type: integer
      required:
        - riderId
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
        comment:
          type: ["string", "null"]
      required:
        - status
    DuplicateActionBody:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject, continue]
        note:
          type: ["string", "null"]
      required:
        - action
    BulkOrderBody:
      type: object
      properties:
        vendorId:
          type: integer
        orders:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderBody"
      required:
        - vendorId
        - orders
    BulkOrderResult:
      type: object
      properties:
        created:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/BulkOrderError"
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required:
        - created
        - failed
        - errors
        - orders
...
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
        visibility:
          type: string
          enum: [all, internal, vendor, rider]
      required:
        - content
...